factor out vect in mask object serialization
